search enabled

+++ src/posts/views.py
@@ -1,17 +1,21 @@
 from urllib.parse import quote_plus
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.db.models import Q
 from django.shortcuts import render, get_object_or_404, redirect
-
+from django.utils import timezone
 from .forms import PostForm
 from .models import Post
 
 
 def post_create(request):  # CREATE
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         print(form.cleaned_data.get('title'))
         instance.save()
         messages.success(request, "Succesfully created")
@@ -21,8 +25,11 @@ def post_create(request):  # CREATE
     return render(request, "post_form.html", context)
 
 
-def post_detail(request, id):  # RETRIEVE
-    instance = get_object_or_404(Post, id=id)
+def post_detail(request, slug=None):  # RETRIEVE
+    instance = get_object_or_404(Post, slug=slug)
+    if instance.draft or instance.publish > timezone.now().date():
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
     share_string = quote_plus(instance.content)
     context = {
         "title": instance.title,
@@ -33,8 +40,18 @@ def post_detail(request, id):  # RETRIEVE
 
 
 def post_list(request):  # list items
-    queryset_list = Post.objects.all()  # .order_by("-timestamp")
-    paginator = Paginator(queryset_list, 10)  # Show 25 contacts per page
+    today = timezone.now().date()
+    queryset_list = Post.objects.active()  # .order_by("-timestamp")
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = Post.objects.all()
+
+    query = request.GET.get("q")
+    if query:
+        queryset_list = queryset_list.filter(
+            Q(title__icontains=query) | Q(content__icontains=query) |
+            Q(user__first_name__icontains=query) |
+            Q(user__last_name__icontains=query))
+    paginator = Paginator(queryset_list, 5)
     page_request_var = "page"
     page = request.GET.get(page_request_var)
     try:
@@ -49,14 +66,17 @@ def post_list(request):  # list items
     context = {
         "object_list": queryset,
         "title": "List",
-        "page_request_var": page_request_var
+        "page_request_var": page_request_var,
+        "today": today,
     }
 
     return render(request, 'post_list.html', context)
 
 
-def post_update(request, id=None):  # UPDATE
-    instance = get_object_or_404(Post, id=id)
+def post_update(request, slug=None):  # UPDATE
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     form = PostForm(
         request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -69,8 +89,10 @@ def post_update(request, id=None):  # UPDATE
     return render(request, 'post_form.html', context)
 
 
-def post_delete(request, id=None):  # DELETE
-    instance = get_object_or_404(Post, id=id)
+def post_delete(request, slug=None):  # DELETE
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Succesfully deleted")
     return redirect("posts:list")
